Add options to protoc to import the right packages
As explained in
[srd-d/proteus#41--comment](https://github.com/src-d/proteus/issues/41#issuecomment-272145393)
the problem is that, for timestamp to work corretly, there is some more
options to pass to the `protoc` command.

This options are a set of strings in the following format:

```
M%s=%s
```

Where the first string is the proto file imported, and the second string
needs to be the go package to import in its place. The different options
are separated by `,`.

This set of options need to be separated of the others set of options  by
`,` as well.

Fixes #41