Support alias types
In order to support alias types, a custom declaration needs to be added to
certain fields. Sadly, there is a known limitation: declarations similar to
`type IntList []int` are not yet supported due to limitations on
`gogo/protobuf`.

I think this limitation can be fixed, though it would require a new protobuf
plugin that might not be simple to get done.